Dev_"Refactor code, add missing field"

+++ Dto/Request/VnPayPaymentRequestDto.cs
@@ -7,7 +7,7 @@ namespace Dtos
     public class VnPayPaymentRequestDto
     {
 
-        public uint Amount { get; set; }
+        public ulong Amount { get; set; }
         public DateTime CreateDate { get; set; }
         public string InvoiceId { get; set; }
     }
